jwt-app/requirements.txt
Flask==1.1.1
Flask-Script==2.0.6
SQLAlchemy==1.3.12
psycopg2==2.8.4
pyjwt==1.7.1
gunicorn==19.9.0
